Mario-more
Even thought I was not required to do this one, I liked the Idea of improving what I already know how to do. So this is just a more complete version of the other mario code, where I get the input from the user and print a pyramid of "#" accordingly to the input. In this example the program prints two pyramids with a space betweem them, one is right hand aligned and the other is left hand aligned.